+++ zcash_client_backend/src/proto.rs
@@ -780,8 +780,7 @@ impl proposal::Proposal {
                                                             PoolType::TRANSPARENT,
                                                             out.index,
                                                         )
-                                                    })?
-                                                    .into_wallet_output(),
+                                                    })?,
                                             );
                                         }
                                     }